[FIX] website_sale: ensure state is in the correct country
Versions
--------
- 16.0+

Steps
-----
1. Enable Stripe Express Checkout;
2. go to eCommerce as Public User;
3. pay via express checkout for a deliverable item;
4. in Stripe, set shipping address to somewhere in California;
5. finish payment.

Issue
-----
The new partner is created with its state set to Cadiz instead of
California.

Cause
-----
When searching for a state using its code, it fetches the first matching
item. The problem is that unlike country codes, state codes aren't
unique, e.g. there are 4 states that match the 'CA' code.

Solution
--------
Filter the state from the country record's `state_ids`.

opw-4396024

X-original-commit: ae8a5053f58dc24d9b1784d99696365bd0ef451e
